pyproject.toml
[tool.poetry]
name = "multiagents"
version = "0.1.0"
description = ""
authors = ["jahco <h@elcorderista>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^6.0.0"
langchain = "^0.3.15"
langgraph = "^0.2.67"
openai = "^1.60.2"
langchain-openai = "^0.3.2"
grandalf = "^0.8"
ipython = "^8.31.0"
langchain-community = "^0.3.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
